requirements.txt
openai==1.90.0
langchain==0.2.*
langchain-community
streamlit
python-dotenv
PyPDF2
faiss-cpu
tiktoken
